// src/app/shared/components/profile-card/model/employee.model.ts
export class Employee {
    ID?: number;
    Picture!: string;
    DisplayName!: string;
    AssignedTasks!: number;
    Position!: string;
    FirstName?: string;
    LastName?: string;
    PhoneNumber?: string;
    Prefix?: string;
    BirthDate?: Date;
    HireDate?: Date;
    Notes?: string;
    Address?: string;
}